fix the random issue
