+++ src/app/components/RandomRecipe/page.js
@@ -1,15 +1,31 @@
-'use client';
+"use client";
 
-import '../../page.css';
+import { useState, useEffect } from "react";
+import { options } from "../../services/Spoonacular";
+import { SingleRecipe } from "../SingleRecipe/SingleRecipe";
 
-// import { useSearchParams } from "next/navigation";
-
-import {SingleRecipe} from './SingleRecipe';
+// import Link from "next/link";
 
 export default function Home() {
+  const [recipeID, setRecipeID] = useState();
+  const fetchRecipeID = async () => {
+    const response = await fetch(
+      `https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/random?&number=1`,
+      options
+    );
+    const data = await response.json();
+    const recipe = data.recipes;
+    const singleRecipe = recipe[0];
+    console.log(singleRecipe.id);
+    setRecipeID(singleRecipe.id);
+  };
+
+  useEffect(() => {
+    fetchRecipeID();
+  }, []);
   return (
     <>
-      <SingleRecipe />
+      <SingleRecipe dishId={recipeID} />
     </>
   );
 }
